requirements/local_debian_requirements.txt

anyio==3.6.1
async-timeout==4.0.2
click==8.1.3


fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
inflection==0.5.1
loguru==0.6.0
orjson==3.7.11
pydantic==1.9.2
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
